chore: logs page tweak
